Add Support for RawBody and Error
Customizing the templates a little bit to add `RawBody()` and `Error()`
methods. This will make it easier for downstream consumers to create
interfaces and to surface the error from the API.

`RawBody` simple returns the Body field, only exposed as a method for
interfaces. `Error` returns the first ErrorModel/IcebergErrorResponse
object in the set of possible responses that is not null and
successfully casts.